OP-465 import logic extracted to multiple files along with flake syntax changes
