+++ web/.jenkins-import/docker-compose.yml
@@ -27,5 +27,4 @@ services:
               && python manage.py migrate \
               && export DJANGO_SETTINGS_MODULE=tellus.settings \
               && python importer.py\
-              && python import_validator.py\
-              && python insert_data.py"
+              && python import_validator.py
